Early resolution of property-placeholder value
This commit fixes an issue, when in XML defined
config there is <context:property-resolver> with
the value if form of value="${placeholder}" thus
it should be resolved with system property. Issue
occurs when property already resolved, but appropriate
splitting by comma already had passed.

Resolution of that property should be on early stage
when BeanDefinition creation is still in progress.
Resolution should be right before execution of
commaDelimitedListToStringArray method.

Issue: SPR-10502